backend/requirements.txt
Django>=5.2,<6
djangorestframework>=3.14
django-cors-headers>=4.0
numpy>=1.26
scikit-learn>=1.3
joblib>=1.3
gunicorn>=21.0
psycopg2-binary
